requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gTTS==2.5.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.8.2
openai==1.40.6
pillow==10.3.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
regex==2024.7.24
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.2
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
